% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito_df.R
\name{mito_df}
\alias{mito_df}
\title{Create mitochondrial loci dataframe}
\usage{
mito_df()
}
\value{
loci dataframe containing the human mitochondrial loci details
}
\description{
Creates the basic dataframe with the human mitochondrial loci,
that will be used to plot the mitochondrial genome using \code{\link{mito_plot}}.
}
