% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mito_plot.R
\name{mito_plot}
\alias{mito_plot}
\title{Plot mitochondrial genome}
\usage{
mito_plot(show_loci_names = TRUE, show_loci_legend = TRUE)
}
\arguments{
\item{show_loci_names}{Show labels of loci names on the genome (default = TRUE)}

\item{show_loci_legend}{Show legend for locus type colours (default = TRUE)}
}
\value{
p ggplot2 polar plot showing the human mitochondrial genome
}
\description{
Creates the basic graphic representation of the human mitochondrial
genome, that will be used to plot variants through \code{\link{plot_df}},
\code{\link{plot_vcf}} or \code{\link{plot_json}}.
}
