% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_json.R
\name{plot_json}
\alias{plot_json}
\title{Plot variants from a JSON file}
\usage{
plot_json(json_file, json_format = "vector", var_col = "POS",
  show_loci_names = TRUE, show_loci_legend = TRUE, title = "",
  save_plot = FALSE, save_to = "mitoviz_plot.png")
}
\arguments{
\item{json_file}{Path of the input JSON file}

\item{json_format}{JSON format used; allowed values are c("vector", "dataframe") (default = "vector")}

\item{var_col}{When json_format = "dataframe", name of column containing variants position (default = "POS")}

\item{show_loci_names}{Show labels of loci names on the genome (default = TRUE)}

\item{show_loci_legend}{Show legend for locus type colours (default = TRUE)}

\item{title}{Set plot title (default = "")}

\item{save_plot}{Save the current plot to a file (default = FALSE)}

\item{save_to}{Path or filename to which the current plot will be saved (default = "mitoviz_plot.png")}
}
\value{
p ggplot2 polar plot variants on the human mitochondrial genome
}
\description{
Adds a geom_point layer to the output of \code{\link{mito_plot}} showing
mitochondrial variants contained in the given JSON file
}
